Refactor: 이메일 워커 설정/병렬 처리 개선
