+++ app/src/main/java/party/manitto/config/SchedulingConfig.kt
@@ -1,8 +1,10 @@
 package party.manitto.config
 
 import org.springframework.context.annotation.Configuration
+import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.scheduling.annotation.EnableScheduling
 
 @Configuration
 @EnableScheduling
+@EnableConfigurationProperties(EmailWorkerProperties::class)
 class SchedulingConfig
